chore: Reorganized tests (docs-for-embroider-css-modules-temporary)
