+++ docs/embroider-css-modules-temporary/package.json
@@ -52,6 +52,8 @@
     "@glint/environment-ember-loose": "^1.0.2",
     "@glint/environment-ember-template-imports": "^1.0.2",
     "@glint/template": "^1.0.2",
+    "@percy/cli": "^1.26.0",
+    "@percy/ember": "^4.2.0",
     "@shared-configs/ember-template-lint": "workspace:*",
     "@shared-configs/eslint-config-ember": "workspace:*",
     "@shared-configs/prettier": "workspace:*",
