+++ docs/embroider-css-modules-temporary/tests/acceptance/index/accessibility-test.ts
@@ -1,0 +1,15 @@
+import { visit } from '@ember/test-helpers';
+import { module, test } from 'qunit';
+
+import { a11yAudit, setupApplicationTest } from '../../helpers';
+
+module('Acceptance | index', function (hooks) {
+  setupApplicationTest(hooks);
+
+  test('Accessibility', async function (assert) {
+    await visit('/');
+    await a11yAudit();
+
+    assert.ok(true);
+  });
+});
