+++ docs/embroider-css-modules-temporary/tests/acceptance/index/as-user-test.ts
@@ -1,16 +1,13 @@
 import { visit } from '@ember/test-helpers';
-import { a11yAudit } from 'ember-a11y-testing/test-support';
 import { module, test } from 'qunit';
 
-import { setupApplicationTest } from '..//helpers';
+import { setupApplicationTest } from '../../helpers';
 
 module('Acceptance | index', function (hooks) {
   setupApplicationTest(hooks);
 
-  test('it renders', async function (assert) {
+  test('We can visit the page', async function (assert) {
     await visit('/');
-
-    assert.ok(true, 'We visited the index route.');
 
     assert.dom('[data-test-local-class-attribute]').hasStyle(
       {
@@ -32,9 +29,5 @@ module('Acceptance | index', function (hooks) {
       },
       'We can use the {{local-class-new}} helper from embroider-css-modules-temporary.',
     );
-
-    await a11yAudit();
-
-    assert.ok(true, 'We passed Axe tests.');
   });
 });
