+++ docs/embroider-css-modules-temporary/tests/acceptance/index/visual-regression-test.ts
@@ -1,0 +1,17 @@
+import { visit } from '@ember/test-helpers';
+import { module, test } from 'qunit';
+
+import { percySnapshot, setupApplicationTest } from '../../helpers';
+
+module('Acceptance | index', function (hooks) {
+  setupApplicationTest(hooks);
+
+  test('Visual regression', async function (assert) {
+    assert.expect(1);
+
+    await visit('/');
+    await percySnapshot(assert);
+
+    assert.ok(true);
+  });
+});
